+++ src/main/java/pony/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package pony.command;
+
+import pony.Storage;
+import pony.task.TaskList;
+import pony.Ui;
+
 public class ExitCommand extends Command {
 
     public ExitCommand() {
