+++ src/main/java/pony/task/Event.java
@@ -1,3 +1,5 @@
+package pony.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 public class Event extends Task {
